requirements.txt
# Astro runtime dependencies. Install from this directory (the Astro project root):
#
#   pip install -r requirements.txt
#   pip install -e .
#
# The installable package lives in the nested ./astro/ subfolder (import name: astro).
#
# Optional training (PyTorch): pip install -r requirements-train.txt

# --- core (matches pyproject [project] dependencies) ---
typing-extensions>=4.8.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
pydantic>=2.0
rank-bm25>=0.2.2
stockstats>=0.6.2
pyarrow>=14.0.0
requests>=2.28.0

# --- API server ---
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# --- IBKR (Gateway / TWS) ---
ib_async>=2.0.0

# --- env files (OpenAI / ASTRO_API_KEY / etc.) ---
python-dotenv>=1.0.0
